requirements.txt
discord.py
python-dotenv
folium
requests